# Move SliceNextElem to internal package and reuse it.