--- a/internal/util.go
+++ b/internal/util.go
@@ -1,0 +1,16 @@
+package internal
+
+import "reflect"
+
+func SliceNextElem(v reflect.Value) reflect.Value {
+	elemType := v.Type().Elem()
+
+	if elemType.Kind() == reflect.Ptr {
+		elem := reflect.New(elemType.Elem())
+		v.Set(reflect.Append(v, elem))
+		return elem.Elem()
+	}
+
+	v.Set(reflect.Append(v, reflect.Zero(elemType)))
+	return v.Index(v.Len() - 1)
+}