```
benchmark                            old ns/op     new ns/op     delta
BenchmarkModelHasOneGopg-4           314800        282860        -10.15%
BenchmarkModelHasManyGopg-4          3166363       3199460       +1.05%
BenchmarkModelHasMany2ManyGopg-4     3557399       3453010       -2.93%

benchmark                            old allocs     new allocs     delta
BenchmarkModelHasOneGopg-4           1173           1081           -7.84%
BenchmarkModelHasManyGopg-4          6596           6504           -1.39%
BenchmarkModelHasMany2ManyGopg-4     8619           8527           -1.07%

benchmark                            old bytes     new bytes     delta
BenchmarkModelHasOneGopg-4           77573         62844         -18.99%
BenchmarkModelHasManyGopg-4          390383        375674        -3.77%
BenchmarkModelHasMany2ManyGopg-4     508625        493882        -2.90%
```